# www/api/openapi.yaml
openapi: 3.0.3
info:
  title: RootWarden API
  description: |
    API REST du backend Flask de RootWarden — plateforme de gestion centralisee
    de serveurs Linux (cles SSH, mises a jour, iptables, scan CVE).

    **Authentification** : header `X-API-KEY` obligatoire sur toutes les routes (sauf /test).
    Le frontend PHP utilise `/api_proxy.php` qui injecte la cle automatiquement.
  version: 1.11.0
  contact:
    name: RootWarden
    url: https://github.com/Timikana/rootwarden
servers:
  - url: https://localhost:5000
    description: Backend direct (dev)
  - url: /api_proxy.php
    description: Via proxy PHP (production)
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
...
security:
  - ApiKeyAuth: []
paths:
...
  /iptables:
    post:
      tags: [Iptables]
      summary: Charger les regles iptables d'un serveur
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [server_ip, ssh_user, action]
              properties:
                server_ip: { type: string }
                ssh_user: { type: string }
                ssh_password: { type: string }
                root_password: { type: string }
                action: { type: string, enum: [get] }
      responses:
        '200': { description: Regles IPv4+IPv6 }
  /iptables-validate:
    post:
      tags: [Iptables]
      summary: Valide les regles iptables (dry-run iptables-restore --test)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [server_ip, ssh_user, rules_v4]
              properties:
                server_ip: { type: string }
                ssh_user: { type: string }
                rules_v4: { type: string }
                rules_v6: { type: string }
      responses:
        '200': { description: Validation OK ou erreur syntaxe }
  /iptables-apply:
    post:
      tags: [Iptables]
      summary: Applique les regles iptables sur un serveur
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [server_ip, ssh_user, rules_v4]
              properties:
                server_ip: { type: string }
                ssh_user: { type: string }
                rules_v4: { type: string }
                rules_v6: { type: string }
                change_reason: { type: string }
      responses:
        '200': { description: Regles appliquees }
  /iptables-restore:
    post:
      tags: [Iptables]
      summary: Restaure les regles iptables depuis la BDD
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [server_ip, ssh_user]
              properties:
                server_ip: { type: string }
                ssh_user: { type: string }
      responses:
        '200': { description: Regles restaurees }
  /iptables-history:
    get:
      tags: [Iptables]
      summary: Historique des modifications iptables
      parameters:
        - { name: server_id, in: query, required: true, schema: { type: integer } }
      responses:
        '200': { description: Liste historique (20 dernieres) }
  /iptables-rollback:
    post:
      tags: [Iptables]
      summary: Rollback vers une version anterieure des regles
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [history_id]
              properties:
                history_id: { type: integer }
      responses:
        '200': { description: Rollback effectue }

  # ── CVE ───────────────────────────────────────────────────────────────
...